refactor: 반복되는 공통 코드 base activity로 이동
