+++ android/app/src/main/java/poke/rogue/helper/presentation/base/BindingActivity.kt
@@ -1,10 +1,19 @@
 package poke.rogue.helper.presentation.base
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
 import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
+import poke.rogue.helper.R
+import poke.rogue.helper.presentation.util.context.drawableOf
+import poke.rogue.helper.presentation.util.context.stringOf
+import poke.rogue.helper.presentation.util.context.toast
 
 abstract class BindingActivity<T : ViewDataBinding>(
     @LayoutRes private val layoutRes: Int,
@@ -14,5 +23,45 @@ abstract class BindingActivity<T : ViewDataBinding>(
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = DataBindingUtil.setContentView(this, layoutRes)
+        initToolbar()
+    }
+
+    abstract fun toolBar(): Toolbar?
+
+    private fun initToolbar() {
+        toolBar()?.let {
+            setSupportActionBar(it)
+            it.overflowIcon = drawableOf(R.drawable.ic_menu)
+            supportActionBar?.setDisplayShowTitleEnabled(true)
+        }
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.menu_toolbar, menu)
+        return super.onCreateOptionsMenu(menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.item_toolbar_pokerogue -> {
+                navigateToPokeRogue()
+            }
+
+            R.id.item_toolbar_feedback -> {
+                toast(R.string.toolbar_feedback)
+            }
+
+            android.R.id.home -> {
+                onBackPressedDispatcher.onBackPressed()
+            }
+        }
+        return true
+    }
+
+    private fun navigateToPokeRogue() {
+        Intent(
+            Intent.ACTION_VIEW,
+            Uri.parse(stringOf(R.string.home_pokerogue_url)),
+        ).also { startActivity(it) }
     }
 }
