+++ android/app/src/main/java/poke/rogue/helper/presentation/home/HomeActivity.kt
@@ -3,9 +3,8 @@ package poke.rogue.helper.presentation.home
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
 import androidx.activity.viewModels
+import androidx.appcompat.widget.Toolbar
 import poke.rogue.helper.R
 import poke.rogue.helper.databinding.ActivityHomeBinding
 import poke.rogue.helper.presentation.ability.AbilityActivity
@@ -13,9 +12,7 @@ import poke.rogue.helper.presentation.base.BindingActivity
 import poke.rogue.helper.presentation.dex.PokemonActivity
 import poke.rogue.helper.presentation.tip.TipActivity
 import poke.rogue.helper.presentation.type.TypeActivity
-import poke.rogue.helper.presentation.util.context.drawableOf
 import poke.rogue.helper.presentation.util.context.stringOf
-import poke.rogue.helper.presentation.util.context.toast
 import poke.rogue.helper.presentation.util.repeatOnStarted
 
 class HomeActivity : BindingActivity<ActivityHomeBinding>(R.layout.activity_home) {
@@ -23,34 +20,14 @@ class HomeActivity : BindingActivity<ActivityHomeBinding>(R.layout.activity_home
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setSupportActionBar(binding.toolbarHome.toolbar)
-        supportActionBar?.setDisplayShowTitleEnabled(false)
         initViews()
         initObservers()
     }
 
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.menu_toolbar, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.item_toolbar_pokerogue -> {
-                navigateToPokeRogue()
-            }
-
-            R.id.item_toolbar_feedback -> {
-                toast(R.string.toolbar_feedback)
-            }
-        }
-        return true
-    }
+    override fun toolBar(): Toolbar = binding.toolbarHome
 
     private fun initViews() =
         with(binding) {
-            setSupportActionBar(toolbarHome.toolbar)
-            toolbarHome.toolbar.overflowIcon = drawableOf(R.drawable.ic_menu)
             supportActionBar?.setDisplayShowTitleEnabled(false)
             supportActionBar?.setDisplayHomeAsUpEnabled(false)
             actionHandler = viewModel
