+++ android/app/src/main/java/poke/rogue/helper/presentation/type/TypeActivity.kt
@@ -2,11 +2,9 @@ package poke.rogue.helper.presentation.type
 
 import android.content.Context
 import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
 import androidx.activity.viewModels
+import androidx.appcompat.widget.Toolbar
 import poke.rogue.helper.R
 import poke.rogue.helper.databinding.ActivityTypeBinding
 import poke.rogue.helper.presentation.base.BindingActivity
@@ -14,8 +12,6 @@ import poke.rogue.helper.presentation.type.model.SelectorType
 import poke.rogue.helper.presentation.type.model.TypeUiModel
 import poke.rogue.helper.presentation.type.result.TypeResultAdapter
 import poke.rogue.helper.presentation.type.selection.TypeSelectionBottomSheetFragment
-import poke.rogue.helper.presentation.util.context.drawableOf
-import poke.rogue.helper.presentation.util.context.stringOf
 import poke.rogue.helper.presentation.util.repeatOnStarted
 
 class TypeActivity : BindingActivity<ActivityTypeBinding>(R.layout.activity_type) {
@@ -24,32 +20,6 @@ class TypeActivity : BindingActivity<ActivityTypeBinding>(R.layout.activity_type
     }
     private val typeResultAdapter by lazy { TypeResultAdapter() }
 
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.menu_toolbar, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.item_toolbar_pokerogue -> {
-                navigateToPokeRogue()
-            }
-
-            R.id.item_toolbar_feedback -> {
-            }
-
-            android.R.id.home -> {
-                finish()
-            }
-        }
-        return true
-    }
-
-    private fun navigateToPokeRogue() {
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse(stringOf(R.string.home_pokerogue_url)))
-        startActivity(intent)
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         initViews()
@@ -57,12 +27,10 @@ class TypeActivity : BindingActivity<ActivityTypeBinding>(R.layout.activity_type
         initObserver()
     }
 
+    override fun toolBar(): Toolbar = binding.toolbarType
+
     private fun initViews() =
         with(binding) {
-            setSupportActionBar(toolbarHome.toolbar)
-            toolbarHome.toolbar.overflowIcon = drawableOf(R.drawable.ic_menu)
-            supportActionBar?.setDisplayShowTitleEnabled(false)
-
             typeHandler = viewModel
             vm = viewModel
             lifecycleOwner = this@TypeActivity
